+++ controllers/contacts.js
@@ -1,0 +1,89 @@
+const Joi = require("joi");
+
+const contacts = require("../../models/contacts");
+
+const { HttpError } = require("../../helpers");
+
+const addSchema = Joi.object({
+  name: Joi.string().required(),
+  email: Joi.string().required(),
+  phone: Joi.string().required(),
+});
+
+const listContacts = async (req, res, next) => {
+  try {
+    const contacts = await contacts.listContacts();
+    if (!contacts) {
+      throw HttpError(404, error.message);
+    }
+    res.json(contacts);
+  } catch (error) {
+    next(error);
+  }
+};
+
+const getContactById = async (req, res, next) => {
+  try {
+    const { contactId } = req.params;
+    const contact = await contacts.getContactById(contactId);
+    if (!contact) {
+      throw HttpError(404, error.message);
+    }
+    res.json(contact);
+  } catch (error) {
+    next(error);
+  }
+};
+
+const addContact = async (req, res, next) => {
+  try {
+    const { error } = addSchema.validate(req.body);
+    if (error) {
+      throw HttpError(400, error.message);
+    }
+    const newContact = await contacts.addContact(req.body);
+    res.status(201).json(newContact);
+  } catch (error) {
+    next(error);
+  }
+};
+
+const removeContact = async (req, res, next) => {
+  try {
+    const { contactId } = req.params;
+    const result = await contacts.removeContact(contactId);
+    if (!result) {
+      throw HttpError(404, error.message);
+    }
+    res.json({
+      message: "Delete success",
+    });
+  } catch (error) {
+    next(error);
+  }
+};
+
+const updateContactById = async (req, res, next) => {
+  try {
+    const { error } = addSchema.validate(req.body);
+    if (error) {
+      throw HttpError(400, error.message);
+    }
+    const { contactId } = req.params;
+    const result = await contacts.updateContactById(contactId, req.body);
+    if (!result) {
+      throw HttpError(404, error.message);
+    }
+    res.json(result);
+  } catch (error) {
+    next(error);
+  }
+};
+
+module.exports = {
+  listContacts,
+  getContactById,
+  addContact,
+  removeContact,
+  updateContactById,
+};
